Divided methods in separeted files
